Code refectoring
