+++ src/components/index.js
@@ -1,2 +1,4 @@
 export {LogoComponent} from './logoComponent';
 export {ScreenContainer} from './ScreenContainer';
+export {RoundProfileImage} from './RoundProfileImage';
+export {SocialPlatforms} from './SocialMediaIcons';
